requirements.txt
Django>=5.2,<6.0
crispy-bootstrap5>=2024.2
Pillow>=10.0.0
django-crispy-forms>=2.0
gunicorn>=21.2.0
whitenoise>=6.6.0
dj-database-url>=2.1.0
psycopg2-binary>=2.9.9
python-decouple>=3.8
